washdb-bot/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "washdb-bot"
version = "0.2.0"
description = "Multi-source web scraping tool for business data discovery and enrichment"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "WashDB Team"}
]
keywords = ["web-scraping", "data-collection", "business-intelligence", "playwright", "beautifulsoup"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "sqlalchemy>=2.0.0,<3.0.0",
    "psycopg[binary]>=3.0.0,<4.0.0",
    "requests>=2.31.0,<3.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.0",
    "python-dotenv>=1.0.0",
    "tldextract>=5.0.0",
    "APScheduler>=3.10.0,<4.0.0",
    "nicegui>=1.4.0",
    "flask>=2.3.0,<4.0.0",
    "flask-cors>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
washdb-gui = "niceui.main:main"

[tool.setuptools]
packages = ["db", "scrape_yp", "scrape_site", "scrape_google", "scrape_bing", "scheduler", "runner", "niceui", "seo_intelligence", "tests"]

# ==========================================
# Black Configuration (Code Formatting)
# ==========================================
[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | __pycache__
  | \.pytest_cache
  | legacy
  | data
  | logs
)/
'''

# ==========================================
# Ruff Configuration (Linting)
# ==========================================
[tool.ruff]
# Enable pycodestyle (E), pyflakes (F), isort (I), and more
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]

# Ignore specific rules
ignore = [
    "E501",   # line too long (handled by black)
]

# Exclude directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    "build",
    "dist",
    "legacy",
    "data",
    "logs",
]

line-length = 100
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = ["S101"]  # Allow asserts in tests

[tool.ruff.isort]
known-first-party = ["scrape_yp", "scrape_google", "scrape_bing", "scrape_site", "niceui", "db", "runner", "scheduler", "seo_intelligence"]

# ==========================================
# Pytest Configuration
# ==========================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",  # Show summary of all test outcomes
    "--strict-markers",
    "--tb=short",
]

# Test markers
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "acceptance: marks tests as end-to-end acceptance tests",
    "network: marks tests that require network access",
]

# Coverage settings
[tool.coverage.run]
source = ["scrape_yp", "scrape_google", "scrape_bing", "scrape_site", "niceui", "db", "runner", "scheduler"]
omit = [
    "*/tests/*",
    "*/legacy/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
